// frontend/src/api/models/FavoritesCarousel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { FavoritesImage } from './FavoritesImage';

export type FavoritesCarousel = {
    id: number;
    user_id: number;
    name: string;
    order: number;
    created_at: string;
    updated_at: string;
    images?: Array<FavoritesImage>;
};